feat: add server side validation

+++ server/src/blogs/blog.router.ts
@@ -1,6 +1,8 @@
 import express, { Request, Response } from 'express';
 import { BlogService } from './blog.service';
 import { BlogType } from './blog.interface';
+import { validationResult, checkSchema } from 'express-validator';
+import { blogValidationSchema } from '../../validationSchema/blogValidationSchema';
 
 export const blogRouter = express.Router();
 
@@ -19,7 +21,7 @@ blogRouter.get('/', async (_, res: Response) => {
     const blog = await BlogService.getAllBlogs();
     res.status(200).send({ success: true, blog: blog });
   } catch (error) {
-        res.status(500).json({ success: false, message: 'Internal server error' });
+    res.status(500).json({ success: false, message: 'Internal server error' });
   }
 });
 
@@ -31,7 +33,9 @@ blogRouter.get(
       const blog = await BlogService.getBlogById(id);
       res.status(200).send({ success: true, blog: blog });
     } catch (error) {
-         res.status(500).json({ success: false, message: 'Internal server error' });
+      res
+        .status(500)
+        .json({ success: false, message: 'Internal server error' });
     }
   },
 );
@@ -44,41 +48,79 @@ blogRouter.get(
       const blog = await BlogService.getBlogBySlug(slug);
       res.status(200).send({ success: true, blog: blog });
     } catch (error) {
-          res.status(500).json({ success: false, message: 'Internal server error' });
+      res
+        .status(500)
+        .json({ success: false, message: 'Internal server error' });
     }
   },
 );
 
-blogRouter.post('/create', async (req: Request<RequestBody>, res: Response) => {
-  try {
-    const {blogInfo} = req.body;
-    const blog = await BlogService.createBlog(blogInfo);
-    res.status(200).send({ success: true, blog: blog });
-  } catch (error) {
-        res.status(500).json({ success: false, message: 'Internal server error' });
-  }
-});
+blogRouter.post(
+  '/create',
+  checkSchema(blogValidationSchema.createBlogSchema),
+  async (req: Request<RequestBody>, res: Response) => {
+    try {
+      const errors = validationResult(req);
+
+      if (!errors.isEmpty()) {
+        return res.status(422).json({
+          errors: errors.array(),
+        });
+      }
+      const blogInfo = req.body;
+      const blog = await BlogService.createBlog(blogInfo);
+      res.status(200).send({ success: true, blog: blog });
+    } catch (error) {
+      res
+        .status(500)
+        .json({ success: false, message: 'Internal server error' });
+    }
+  },
+);
 
 blogRouter.post(
   '/update/:id',
+  checkSchema(blogValidationSchema.createBlogSchema),
   async (req: Request<RequestParams, RequestBody>, res: Response) => {
     try {
-      const {blogInfo} = req.body;
-      const {id} = req.params;
+      const errors = validationResult(req);
+
+      if (!errors.isEmpty()) {
+        return res.status(422).json({
+          errors: errors.array(),
+        });
+      }
+      const blogInfo = req.body;
+      const { id } = req.params;
       const blog = await BlogService.updateBlog(id, blogInfo);
       res.status(200).send({ success: true, blog: blog });
     } catch (error) {
-          res.status(500).json({ success: false, message: 'Internal server error' });
+      res
+        .status(500)
+        .json({ success: false, message: 'Internal server error' });
     }
   },
 );
 
-blogRouter.post('/delete', async (req: Request<RequestBody>, res: Response) => {
-  try {
-    const { id } = req.body;
-    const blog = await BlogService.deleteBlog(id);
-    res.status(200).send({ success: true, blog: blog });
-  } catch (error) {
-        res.status(500).json({ success: false, message: 'Internal server error' });
-  }
-});
+blogRouter.post(
+  '/delete',
+  checkSchema(blogValidationSchema.deleteSchema),
+  async (req: Request<RequestBody>, res: Response) => {
+    try {
+      const errors = validationResult(req);
+
+      if (!errors.isEmpty()) {
+        return res.status(422).json({
+          errors: errors.array(),
+        });
+      }
+      const { id } = req.body;
+      const blog = await BlogService.deleteBlog(id);
+      res.status(200).send({ success: true, blog: blog });
+    } catch (error) {
+      res
+        .status(500)
+        .json({ success: false, message: 'Internal server error' });
+    }
+  },
+);
